Move input and section components to folder
