+++ src/shared/components/inputs/TextInput.tsx
@@ -1,5 +1,5 @@
 import { type InputHTMLAttributes } from 'react';
-import Input from './Input';
+import Input from '../Input';
 
 interface TextInputProps extends InputHTMLAttributes<HTMLInputElement> {
   label: string;
