+++ src/shared/components/sections/SectionHeader.tsx
